fix: category page latests posts
