+++ packages/blog/src/components/category/CategoryLatestPosts/index.ts
@@ -1,2 +1,1 @@
-export * from './CategoryPage'
 export * from './CategoryLatestPosts'
