+++ packages/blog/src/components/home/HomeLatestPosts/index.tsx
@@ -1,0 +1,1 @@
+export * from './HomeLatestPosts'
