+++ packages/blog/src/components/home/HomePage/HomePage.tsx
@@ -1,8 +1,9 @@
 import parseISO from 'date-fns/parseISO'
 import { useTranslation } from 'next-i18next'
 
-import { Metadata, HeaderImage, LatestPosts } from '@components/common'
+import { Metadata, HeaderImage } from '@components/common'
 import { useGetCategories } from '@hooks'
+import { HomeLatestPosts } from '../HomeLatestPosts'
 
 export interface HomePageProps {
   latestPosts: Post[]
@@ -30,7 +31,7 @@ export function HomePage({ latestPosts }: HomePageProps) {
       ) : (
         <div>{t('homePage.noPublishedArticles')}</div>
       )}
-      <LatestPosts title={t('latestPosts.title')} categories={categories} posts={latestPosts} />
+      <HomeLatestPosts title={t('latestPosts.title')} categories={categories} posts={latestPosts} />
     </>
   )
 }
